Avoid deeply nesting operators when unnecessary
Takes advantage of regex rules to simplify applied occurrences of
repetition and option operators.

Unfortunately, this cannot be propagated to the Option and Repetition
constructors because they are case classes, and the magic `this` method
must delegate to other magic `this` methods. That is, I can't have
`Option(Repetition('A))` return `Repetition('A)`.